add: initial solution to scalar

+++ scalar/src/lib.rs
@@ -1,21 +1,20 @@
-pub fn sum(a: u8, b: u8) -> u16 {
-    a as u16 + b as u16
+pub fn sum(a: u8, b: u8) -> u8 {
+    a + b
 }
 
 pub fn diff(a: i16, b: i16) -> i16 {
     a - b
 }
 
-pub fn pro(a: i8, b: i8) -> i16 {
-    a as i16 * b as i16
+pub fn pro(a: i8, b: i8) -> i8 {
+    a * b
 }
 
-pub fn quo(a: i32, b: i32) -> i32 {
+pub fn quo(a: f64, b: f64) -> f64 {
     a / b
-
 }
 
-pub fn rem(a: i32, b: i32) -> i32 {
+pub fn rem(a: f64, b: f64) -> f64 {
     a % b
 }
